Moves `FeedImageDataLoader` and `FeedImageDataLoaderTask` protocols to separate file within corresponding folder.
